# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Arceus Internal MCP API
  version: "1.0.0"
  summary: Fastify-backed internal control plane for agent heartbeats.
  description: |
    Headless API consumed by Arceus agents. All routes live under
    `/api/internal/v1/*`, require bearer auth, and accept an
    `Idempotency-Key` header for safe replay.
servers:
  - url: http://localhost:3001
    description: Local development
  - url: https://arceus.internal
    description: Internal staging
security:
  - BearerAuth: []
...
paths:
...
  /api/internal/v1/artifacts:
    post:
      tags: [artifacts]
      summary: Create an artifact
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateArtifactBody" }
      responses:
        "201":
          description: Artifact created
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArtifactCreatedEnvelope" }
        "422": { $ref: "#/components/responses/Validation" }
...
  /api/internal/v1/approvals:
    post:
      tags: [approvals]
      summary: Request an approval
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateApprovalBody" }
      responses:
        "201":
          description: Approval requested
          headers:
            Location:
              schema: { type: string }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Validation" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Conflict:
      description: Idempotency conflict or invalid state
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Validation:
      description: Request body failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationErrorEnvelope" }
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [status, summary]
      properties:
        status: { type: string, enum: [success] }
        summary: { type: string }
        data: { type: object, additionalProperties: true }
        nextActions:
          type: array
          items: { type: string }
...
    ErrorEnvelope:
      type: object
      required: [status, summary, error]
      properties:
        status: { type: string, enum: [error] }
        summary: { type: string }
        error:
          type: object
          required: [cause, retry, stopWhen]
          properties:
            cause:
              type: string
              enum: [validation, governance, not_found, conflict, upstream, internal]
            retry:
              type: string
              enum: [safe, unsafe, never]
            stopWhen: { type: string }
    ValidationErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/ErrorEnvelope"
        - type: object
          properties:
            error:
              type: object
              properties:
                details:
                  type: array
                  items:
                    type: object
                    properties:
                      field: { type: string }
                      message: { type: string }
                      code: { type: string }
...
    CreateArtifactBody:
      type: object
      required: [agent, kind, title, content]
      properties:
        agent: { type: string, maxLength: 100 }
        kind:
          type: string
          enum: [plan, code, output, specification]
        title: { type: string, maxLength: 300 }
        content: { type: string, maxLength: 200000 }
        taskId: { type: string }
    ArtifactCreatedEnvelope:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [artifactId]
              properties:
                artifactId: { type: string }
...
    CreateApprovalBody:
      type: object
      required: [type, requestedByRole, title, description]
      properties:
        type:
          type: string
          enum: [strategy, hire, meeting_blocker, external_action, tool_governance]
          description: The Approval domain type. Agent chooses based on what it is asking approval for.
        requestedByRole:
          type: string
          enum: [ceo, cto, pm, developer, tester, ui_designer, marketing, skills_lead]
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        meetingId:
          type: string
          nullable: true
        agendaItemId:
          type: string
          nullable: true